adds new modules and tests passing
